+++ grakn-engine/src/main/java/ai/grakn/engine/Grakn.java
@@ -19,7 +19,7 @@
 package ai.grakn.engine;
 
 import ai.grakn.GraknSystemProperty;
-import ai.grakn.engine.processid.GraknPid;
+import ai.grakn.engine.grakn_pid.GraknPid;
 import org.hyperic.sigar.Sigar;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
